+++ SimpleTemplate-Shop/Models/CartLine.cs
@@ -8,7 +8,7 @@ namespace SimpleTemplate_Shop.Models
         public int ProductId { get; set; }
         public Product Product { get; set; }
         public int Quantity { get; set; }
-        public decimal TotalSum { get; set; }
+        public int TotalSum { get; set; }
         public DateTime OrderTime { get; set; } = DateTime.Now;
     }
 }
